xg fix

+++ sql/initDB4.sql
@@ -146,7 +146,7 @@ CREATE TABLE Substitutions (
 );
 
 CREATE TABLE xGoals (
-    total_xg INTEGER PRIMARY KEY,
+    total_xg numeric(5,2) PRIMARY KEY,
     matches_played INTEGER,
     FOREIGN KEY (player_id) INTEGER REFERENCES Players(player_id),
 );
